requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
blinker==1.8.2
    # via flask
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
flask==3.0.3
    # via app_buisness (pyproject.toml)
idna==3.7
    # via requests
importlib-metadata==8.4.0
    # via
    #   flask
    #   yoyo-migrations
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
psycopg2==2.9.9
    # via app_buisness (pyproject.toml)
python-dotenv==1.0.1
    # via app_buisness (pyproject.toml)
pyyaml==6.0.2
    # via app_buisness (pyproject.toml)
requests==2.32.3
    # via app_buisness (pyproject.toml)
sqlparse==0.5.1
    # via yoyo-migrations
tabulate==0.9.0
    # via yoyo-migrations
urllib3==2.2.2
    # via requests
werkzeug==3.0.4
    # via flask
yoyo-migrations==9.0.0
    # via app_buisness (pyproject.toml)
zipp==3.20.0
    # via importlib-metadata
pyjwt==2.9.0